requirements.txt
fastapi
uvicorn
pydantic
requests
openai
pytest
python-dotenv
